requirements.txt
django-extensions==2.0.7
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyreadline==2.1
six==1.11.0
typing==3.6.4
